requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
RPi.GPIO>=0.7.1; platform_machine == 'armv7l' or platform_machine == 'aarch64'
